requirements.txt
numpy~=1.19.2
tqdm~=4.50.2
pytorch~=1.7.0